Consistency tests on distributed updates
Enabled with the TEST_CONSISTENCY macro defined.

Requires each thread to keep a copy of the _entire_ distributed matrix.
These are then summed following flush() and the result compared to each
local block. This is clearly very memory intensive and should only be
run on modestly sized matrices.
For the summation to work, MPI must be initialized by the caller (an
assertion will fail in the ConvergentMatrix constructor otherwise).