ensure that MPI is initialized properly if Pthread
per the MPI standard, one cannot use MPI_INIT or
MPI_INIT_THREAD(..,MPI_THREAD_SINGLE,..) if threads are used _anywhere_
in the program.

in practice, this does not matter, thanks to glibc being always
thread-safe today, but it is prudent to hedge against incorrect usage of
MPI in case there are systems that break under the aforementioned
circumstances.

i assume that GASNet sets up MPI's use of threads properly when it uses
MPI+Threads.