Experiment: Use eager transfer of buffered updates
Rather than use view-based serialization when transferring remote update
data, eagerly allocate and transfer data to the remote process.

Since remote allocation is blocking and occurs via RPC, this adds a
naive backpressure mechanism (i.e. the remote process must make user
level progress in order for memory to be allocated, which also implies
that updates are being applied and transfer buffers freed).

This is close to the orginal impl of Bin in upcxx v0.1.